Conflicts:
	announcements/src/org/labkey/announcements/AnnouncementsController.java
	api/src/org/labkey/api/data/TableViewForm.java
	api/src/org/labkey/api/security/permissions/AbstractContainerScopingTest.java
	assay/src/org/labkey/assay/AssayController.java
	core/src/org/labkey/core/CoreController.java
	core/src/org/labkey/core/admin/AdminController.java
	mothership/src/org/labkey/mothership/MothershipController.java
	pipeline/src/org/labkey/pipeline/api/PipelineStatusManager.java
	query/src/org/labkey/query/QueryModule.java
	study/src/org/labkey/study/model/ParticipantGroupManager.java
	studydesign/src/org/labkey/studydesign/StudyDesignController.java
	studydesign/src/org/labkey/studydesign/StudyDesignModule.java
	studydesign/src/org/labkey/studydesign/model/TreatmentManager.java

+++ api/src/org/labkey/api/study/publish/PublishConfirmForm.java
@@ -49,7 +49,7 @@ public class PublishConfirmForm extends ViewForm implements DataRegionSelection.
         return BaseViewAction.springBindParameters(this, "form", pvs);
     }
 
-    private Integer _rowId;
+    private Long _rowId;
     private String[] _targetStudy;
     private String[] _participantId;
     private String[] _visitId;
@@ -64,12 +64,12 @@ public class PublishConfirmForm extends ViewForm implements DataRegionSelection.
     private String _defaultValueSource = DefaultValueSource.PublishSource.name();
     private String _autoLinkCategory;
 
-    public Integer getRowId()
+    public Long getRowId()
     {
         return _rowId;
     }
 
-    public void setRowId(Integer rowId)
+    public void setRowId(Long rowId)
     {
         _rowId = rowId;
     }
